// src/modules/inventory/reports/yarn-store/date-wise-yarn-receive-register-report/date-wise-yarn-receive-register-report-type.ts
export type DateWiseYarnReceiveRegisterReportType = {
  YARN_RECEIVED_DATE: Date;
  SUPPLIER?: string;
  RCVD_CHALLAN_NO?: string;
  YARN_NAME?: string;
  QUANTITY: number;
  LC_UNIT_PRICE: number;
  VALUE: number;
  CURRENCYCODE?: string;
  BBLC_NO?: string;
  BBLC_DATE?: Date;
  WO_NO?: string;
  WO_DATE?: Date;
  COMPANY_NAME?: string;
  COMPANY_ADDRESS?: string;
  COMPANY_REMARKS?: string;
};